refactor: change transaction domain protocol to an specific folder
