+++ src/data/protocols/db/find-transaction-by-filters-repository.ts
@@ -1,5 +1,5 @@
 import { TransactionModel } from "../../../domain/models/transaction";
-import { TransactionFiltersModel } from "src/domain/usecases/find-transaction-by-filters";
+import { TransactionFiltersModel } from "src/domain/usecases/transaction/find-transaction-by-filters";
 
 export interface FindTransactionByFiltersRepository {
   find(filters: TransactionFiltersModel): Promise<TransactionModel[]>;
