+++ src/data/usecases/db-create-transaction.ts
@@ -3,7 +3,7 @@ import { TransactionModel } from "src/domain/models/transaction";
 import {
   CreateTransaction,
   CreateTransactionModel,
-} from "../../domain/usecases/create-transaction";
+} from "../../domain/usecases/transaction/create-transaction";
 import { CreateTransactionRepository } from "../protocols/db/create-transaction-repository";
 
 export class DbCreateTransaction implements CreateTransaction {
