+++ src/data/usecases/db-find-transaction-by-filters.ts
@@ -2,7 +2,7 @@ import { TransactionModel } from "src/domain/models/transaction";
 import {
   FindTransactionByFilters,
   TransactionFiltersModel,
-} from "../../domain/usecases/find-transaction-by-filters";
+} from "../../domain/usecases/transaction/find-transaction-by-filters";
 import { FindTransactionByFiltersRepository } from "../protocols/db/find-transaction-by-filters-repository";
 
 export class DbFindTransactionByFilters implements FindTransactionByFilters {
