+++ src/data/usecases/db-get-total-amount.ts
@@ -1,5 +1,5 @@
 import { TotalAmountModel } from "src/domain/models/total-amount";
-import { GetTotalAmount } from "../../domain/usecases/get-total-amount";
+import { GetTotalAmount } from "../../domain/usecases/transaction/get-total-amount";
 import { GetTotalAmountRepository } from "../protocols/db/get-total-amount-repository";
 
 export class DbGetTotalAmount implements GetTotalAmount {
