+++ src/domain/usecases/transaction/find-transaction-by-id.ts
@@ -1,4 +1,4 @@
-import { TransactionModel } from "../models/transaction";
+import { TransactionModel } from "../../models/transaction";
 
 export interface FindTransactionById {
   findById(_id: string): Promise<TransactionModel>;
