+++ tests/data/usecases/db-create-transaction.spec.ts
@@ -1,7 +1,6 @@
 import { DbCreateTransaction } from "../../../src/data/usecases/db-create-transaction";
 import { CreateTransactionRepository } from "../../../src/data/protocols/db/create-transaction-repository";
 import { TransactionModel } from "../../../src/domain/models/transaction";
-import { CreateTransactionModel } from "../../../src/domain/usecases/create-transaction";
 import crypto from "crypto";
 
 const makeTransactionData = (): {
