+++ tests/presentation/create-transaction-controller.spec.ts
@@ -1,7 +1,7 @@
 import {
   CreateTransaction,
   CreateTransactionModel,
-} from "../../src/domain/usecases/create-transaction";
+} from "../../src/domain/usecases/transaction/create-transaction";
 import { TransactionModel } from "../../src/domain/models/transaction";
 import { CreateTransactionController } from "../../src/presentation/controllers/create-transaction-controller";
 import { Validator } from "../../src/presentation/protocols/validator";
