+++ tests/presentation/index-transaction-controller.spec.ts
@@ -1,7 +1,7 @@
 import {
   FindTransactionByFilters,
   TransactionFiltersModel,
-} from "../../src/domain/usecases/find-transaction-by-filters";
+} from "../../src/domain/usecases/transaction/find-transaction-by-filters";
 import { TransactionModel } from "../../src/domain/models/transaction";
 import { FindTransactionByFiltersController } from "../../src/presentation/controllers/index-transaction-controller";
